refactor(tools): deepen FileWorkspace in tools/fs
Move path resolve, runtime mounts, and file/workspace locks into one
module so filesystem tools are thin adapters over a single seam.
